cv-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.2
pydantic>=2.0
python-dotenv>=1.0.0
loguru>=0.7.0
pillow>=10.0.0
numpy>=1.24.0
aiofiles>=23.2.1
